updated and restructed PHP petstore sample app

+++ samples/petstore/php/petstore/models/Tag.php
@@ -24,12 +24,12 @@
 class Tag {
 
   static $swaggerTypes = array(
-      'id' => 'long',
+      'id' => 'int',
       'name' => 'string'
 
     );
 
-  public $id; // long
+  public $id; // int
   public $name; // string
   }
